LayoutTests/platform/glib/storage/domstorage/sessionstorage/blocked-file-access-expected.txt

CONSOLE MESSAGE: PASS: window.sessionStorage was accessible

Test that we are permitted access to sessionStorage from a file URL if universal access is turned off.
